--- a/src/JustEat.StatsD/Transport/Data.cs
+++ b/src/JustEat.StatsD/Transport/Data.cs
@@ -2,32 +2,11 @@ using System;
 
 namespace JustEat.StatsD
 {
-    public static class DataExtensions
-    {
-        public static ReadOnlySpan<byte> GetSpan(this in Data src)
-        {
-            if (src.Array == null)
-            {
-                return src.Span;
-            }
-            return src.Array.AsSpan();
-        }
-
-        public static byte[] GetArray(this in Data src)
-        {
-            if (src.Array == null)
-            {
-                return src.Span.ToArray();
-            }
-            return src.Array;
-        }
-    }
-
     public ref struct Data
     {
         public Data(byte[] array) : this()
         {
-            Array = array;
+            Array = array ?? throw new ArgumentNullException(nameof(array));
         }
 
         public Data(ReadOnlySpan<byte> span) : this()