--- a/src/JustEat.StatsD/Transport/IStatsDTransport.cs
+++ b/src/JustEat.StatsD/Transport/IStatsDTransport.cs
@@ -2,6 +2,6 @@ namespace JustEat.StatsD
 {
     public interface IStatsDTransport
     {
-        void Send(string metric);
+        void Send(in Data metric);
     }
 }