--- a/src/JustEat.StatsD/Transport/IpTransport.cs
+++ b/src/JustEat.StatsD/Transport/IpTransport.cs
@@ -15,11 +15,6 @@ namespace JustEat.StatsD
 
         public void Send(in Data metric)
         {
-            if (string.IsNullOrWhiteSpace(metric))
-            {
-                return;
-            }
-
             var endpoint = _endpointSource.GetEndpoint();
 
             using (var socket = CreateSocket())