--- a/src/JustEat.StatsD/Transport/UdpTransport.cs
+++ b/src/JustEat.StatsD/Transport/UdpTransport.cs
@@ -1,13 +1,11 @@
 using System;
 using System.Net.Sockets;
-using System.Text;
-#if !NET451
 using System.Runtime.InteropServices;
-#endif
 using JustEat.StatsD.EndpointLookups;
 
 namespace JustEat.StatsD
 {
+
     public class UdpTransport : IStatsDTransport
     {
         private readonly IPEndPointSource _endpointSource;
@@ -17,26 +15,24 @@ namespace JustEat.StatsD
             _endpointSource = endPointSource ?? throw new ArgumentNullException(nameof(endPointSource));
         }
 
-        public void Send(string metric)
+        public void Send(in Data metric)
         {
-            if (string.IsNullOrWhiteSpace(metric))
-            {
-                return;
-            }
-
-            var bytes = Encoding.UTF8.GetBytes(metric);
             var endpoint = _endpointSource.GetEndpoint();
 
             using (var socket = CreateSocket())
             {
-                socket.SendTo(bytes, endpoint);
+#if NETCOREAPP2_1
+                socket.Connect(endpoint);
+                socket.Send(metric.GetSpan());
+#else
+                socket.SendTo(metric.GetArray(), endpoint);
+#endif
             }
         }
 
         private static Socket CreateSocket()
         {
             var socket = new Socket(AddressFamily.InterNetwork, SocketType.Dgram, ProtocolType.Udp);
-
 #if !NET451
             // See https://github.com/dotnet/corefx/pull/17853#issuecomment-291371266
             if (!RuntimeInformation.IsOSPlatform(OSPlatform.OSX))
@@ -46,7 +42,6 @@ namespace JustEat.StatsD
 #else
             socket.SendBufferSize = 0;
 #endif
-
             return socket;
         }
     }