Added multi runtime benchmarks
Added multi runtime tests
Added support for span apis in the main lib without breaking compatibility with 451


Performance runs on my machine looks like this:

| Method | Runtime |     Toolchain |         Mean |      Error |      StdDev |       Median |  Gen 0 | Allocated |
| -------- |-------- |-------------- |-------------:|-----------:|------------:|-------------:|-------:|----------:|
|  RunNull |     Clr |       Default |     2.213 us |  0.0120 us |   0.0172 us |     2.210 us | 0.0496 |     968 B |
|   RunUdp |     Clr |       Default | 1,178.532 us | 68.9897 us | 101.1242 us | 1,139.862 us |      - |    3104 B |
|    RunIp |     Clr |       Default | 1,186.793 us | 68.2845 us |  97.9316 us | 1,128.058 us |      - |    3104 B |
|  RunNull |    Core | .NET Core 2.0 |     2.075 us |  0.0073 us |   0.0102 us |     2.071 us | 0.0420 |     824 B |
|   RunUdp |    Core | .NET Core 2.0 | 1,186.830 us | 45.9153 us |  64.3669 us | 1,175.159 us |      - |    2016 B |
|    RunIp |    Core | .NET Core 2.0 | 1,188.350 us | 35.6655 us |  49.9981 us | 1,176.854 us |      - |    2016 B |
|  RunNull |    Core | .NET Core 2.1 |     1.834 us |  0.1019 us |   0.1493 us |     1.822 us | 0.0267 |     520 B |
|   RunUdp |    Core | .NET Core 2.1 | 1,239.935 us | 46.4202 us |  61.9697 us | 1,214.226 us |      - |    1712 B |
|    RunIp |    Core | .NET Core 2.1 | 1,178.822 us | 29.3535 us |  42.0978 us | 1,183.376 us |      - |    1712 B |

I've changed `IStatsDTransport` interface, not sure if it should be considered as a breaking change.