@AnthonySteele 2018-07-13
With these benchmark stats, you have to make it simple and clear:
what are the timings and memory usage before and after the change?
If i should be comparing .NET Core 2.0 to .NET Core 2.1 then it's much the same?