@dv00d00 2018-07-13
### master branch netcoreapp 2.0

BenchmarkDotNet=v0.10.14, OS=Windows 10.0.17134
Intel Core i5-7200U CPU 2.50GHz (Kaby Lake), 1 CPU, 4 logical and 2 physical cores
  [Host]     : .NET Framework 4.7.1 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.7.3131.0
  Job-FWIARP : .NET Core 2.0.7 (CoreCLR 4.6.26328.01, CoreFX 4.6.26403.03), 64bit RyuJIT

Jit=RyuJit  Runtime=Core  Server=True
Toolchain=.NET Core 2.0

 |Method |     Mean |     Error |    StdDev | Allocated |
|------- |---------:|----------:|----------:|----------:|
 |RunUdp | 1.724 ms | 0.0346 ms | 0.0975 ms |   1.97 KB |
  |RunIp | 1.560 ms | 0.0335 ms | 0.0912 ms |   1.97 KB |

### this branch netcoreapp 2.0

BenchmarkDotNet=v0.10.14, OS=Windows 10.0.17134
Intel Core i5-7200U CPU 2.50GHz (Kaby Lake), 1 CPU, 4 logical and 2 physical cores
.NET Core SDK=2.1.300
  [Host]     : .NET Core 2.0.7 (CoreCLR 4.6.26328.01, CoreFX 4.6.26403.03), 64bit RyuJIT
  Job-YQREIB : .NET Core 2.0.7 (CoreCLR 4.6.26328.01, CoreFX 4.6.26403.03), 64bit RyuJIT

Jit=RyuJit  Runtime=Core  Server=True
Toolchain=.NET Core 2.0

 |Method |     Mean |     Error |    StdDev | Allocated |
|------- |---------:|----------:|----------:|----------:|
 |RunUdp | 1.619 ms | 0.0358 ms | 0.1028 ms |   1.97 KB |
  |RunIp | 1.584 ms | 0.0461 ms | 0.1323 ms |   1.97 KB |

### this branch netcoreapp 2.1

BenchmarkDotNet=v0.10.14, OS=Windows 10.0.17134
Intel Core i5-7200U CPU 2.50GHz (Kaby Lake), 1 CPU, 4 logical and 2 physical cores
  [Host]     : .NET Framework 4.7.1 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.7.3131.0
  Job-JJUQAD : .NET Core 2.1.1 (CoreCLR 4.6.26606.02, CoreFX 4.6.26606.05), 64bit RyuJIT

Jit=RyuJit  Runtime=Core  Server=True
Toolchain=.NET Core 2.1

 |Method |     Mean |     Error |    StdDev | Allocated |
|------- |---------:|----------:|----------:|----------:|
 |RunUdp | 1.516 ms | 0.0303 ms | 0.0568 ms |   1.67 KB |
  |RunIp | 1.770 ms | 0.0698 ms | 0.2047 ms |   1.67 KB |


Timing are basically same, I was getting different results from same code after a minute. Probably due to  antivirus or other apps runing on my laptop.

The benefint as I see it is this pr enables further span based optimizations for non alocating formatter in netcoreapp21 + it removes allocation on from UTF8.GetBytes for short strings