@martincostello 2018-07-13
So UDP is _slightly_ faster and IP is _slightly_ slower.

To me, they're not very compelling changes for such disruptive changes to the public contract and the additional complexity.