@dv00d00 2018-07-13
the code I've used for benchamrks is differs from what is in master

```csharp
    public class FastRunOnMultipleRuntimes : ManualConfig
    {
        public FastRunOnMultipleRuntimes(Job basis)
        {
            Add(basis.With(Runtime.Core).With(Jit.RyuJit).WithGcServer(true).With(CsProjCoreToolchain.NetCoreApp20));
            Add(DefaultConfig.Instance);
            Add(MemoryDiagnoser.Default);
        }
    }

    class Program
    {
        static void Main(string[] args)
        {
            BenchmarkRunner.Run<StatSendingBenchmark>(new FastRunOnMultipleRuntimes(Job.Default));
        }
    }

    [MemoryDiagnoser]
    public class StatSendingBenchmark
    {
        private IStatsDPublisher _udpSender;
        private IStatsDPublisher _ipSender;

        private static readonly TimeSpan Timed = TimeSpan.FromMinutes(1);
        private static readonly DateTime Now = DateTime.Now;

        [GlobalSetup]
        public void Setup()
        {
            var config = new StatsDConfiguration
            {
                // if you want verify that stats are recevied,
                // you will need the IP of suitable local test stats server
                Host = "127.0.0.1",
                Prefix = "testmetric"
            };

            var endpointSource = EndpointParser.MakeEndPointSource(
                config.Host, config.Port, config.DnsLookupInterval);

            var udpTransport = new UdpTransport(endpointSource);
            var ipTransport = new IpTransport(endpointSource);

            _udpSender = new StatsDPublisher(config, udpTransport);
            _udpSender.Increment("startup.ud");

            _ipSender = new StatsDPublisher(config, ipTransport);
            _udpSender.Increment("startup.ip");
        }

        [Benchmark]
        public void RunUdp()
        {
            _udpSender.Increment("increment.ud");
            _udpSender.Timing(Timed, "timer.ud");
            _udpSender.Gauge(long.MaxValue / 2, "gauge.ud", Now);
        }

        [Benchmark]
        public void RunIp()
        {
            _ipSender.Increment("increment.ip");
            _ipSender.Timing(Timed, "timer.ip");
            _ipSender.Gauge(long.MaxValue / 2, "gauge.ip", Now);
        }
    }
```