@dv00d00 2018-07-13
And to provide more context, again, the plan is to add non-allocating formatter + pool udp sockets

So we would get this

netcoreapp2.1
```
.NET Core SDK=2.1.300
  [Host]     : .NET Core 2.1.1 (CoreCLR 4.6.26606.02, CoreFX 4.6.26606.05), 64bit RyuJIT
  Job-EWLOSH : .NET Core 2.1.1 (CoreCLR 4.6.26606.02, CoreFX 4.6.26606.05), 64bit RyuJIT

Jit=RyuJit  Runtime=Core  Server=True
Toolchain=.NET Core 2.1  Categories=UDP
```
|           Method |     Mean |     Error |   StdDev | Allocated |
|----------------- |---------:|----------:|---------:|----------:|
| RunUdpSpanPooled | 38.28 us | 0.8344 us | 1.439 us |       0 B |


Instead of this

```
BenchmarkDotNet=v0.10.14, OS=Windows 10.0.17134
Intel Core i5-7200U CPU 2.50GHz (Kaby Lake), 1 CPU, 4 logical and 2 physical cores
.NET Core SDK=2.1.300
  [Host]     : .NET Core 2.0.7 (CoreCLR 4.6.26328.01, CoreFX 4.6.26403.03), 64bit RyuJIT
  Job-OGUMNE : .NET Core 2.0.7 (CoreCLR 4.6.26328.01, CoreFX 4.6.26403.03), 64bit RyuJIT

Jit=RyuJit  Runtime=Core  Server=True
Toolchain=.NET Core 2.0
```

| Method |     Mean |     Error |   StdDev | Allocated |
|------- |---------:|----------:|---------:|----------:|
| RunUdp | 42.57 us | 0.8365 us | 1.374 us |     816 B |


on a pooled socket transport.

There are 816 bytes allocations from calling 3 methods

```csharp
            _udpSender.Increment("increment.ud");
            _udpSender.Timing(Timed, "timer.ud");
            _udpSender.Gauge(long.MaxValue / 2, "gauge.ud", Now);
```

which could be reduced to 0. Assuming lots of apps are instrumented with statsdpublisher on a middleware level, this could end up in non-trivial gains.