@MuayThai 2018-07-16
I think both of the those tests might be using UDP - "SocketType will sometimes implicitly indicate which ProtocolType will be used within an AddressFamily. For example when the SocketType is Dgram, the ProtocolType is always Udp."
https://msdn.microsoft.com/en-us/library/system.net.sockets.sockettype(v=vs.110).aspx

I think the only difference is the buffer: https://github.com/justeat/JustEat.StatsD/blob/f9dc8f0dda2303df76f83e0309437ef6d0e154c8/src/JustEat.StatsD/UdpTransport.cs#L44