@martincostello 2018-07-18
We've just merged some changes to `master` for what would tentatively be a `3.1.0` release that make [a sizeable dent in](https://github.com/justeat/JustEat.StatsD/pull/81#issuecomment-405935145) the throughput of sending metrics when using the alternate [`PooledUdpTransport`](https://github.com/justeat/JustEat.StatsD/blob/master/src/JustEat.StatsD/PooledUdpTransport.cs) implementation.

We're looking at making this the default (i.e. being the replacement `UdpTransport`) implementation as part of a `4.0.0` release as it would necessitate adding `IDisposable` to some of the types, that would effectively be a breaking change, as otherwise it would create resource leaks.